requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.3
django-jalali==6.0.1
django-jalali-date==1.0.2
django-render-partial==0.4
django-resized==1.0.2
djangorestframework==3.16.0
frozenlist==1.7.0
idna==3.10
jdatetime==4.1.1
multidict==6.6.3
Pillow==10.0.0
propcache==0.3.2
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.0
pytz==2023.3
requests==2.31.0
setuptools==80.9.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
TgCrypto==1.2.5
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.20.1
